Reorganize folder structure
Server side stuff goes under server
Config files go in config
Db directory is still there, but only for db setup, schema stuff
