+++ config/db.json
@@ -1,0 +1,6 @@
+{
+	"host": "localhost",
+	"database": "poemlab",
+	"user": "poemlab",
+	"password": "billygrippa"
+}
