+++ config/security.example.json
