+++ config/security.json
@@ -1,0 +1,3 @@
+{
+  "sessionKey": "That guy with the messed up face"
+}
