+++ server/routes/register.js
@@ -2,7 +2,7 @@ var respond = require('./common').respond;
 var passport = require('passport');
 
 module.exports = function(dbConfig) {
-  var poetsRepo = require("../lib/repositories/poets_repository")(dbConfig);
+  var poetsRepo = require("../repositories/poets_repository")(dbConfig);
 
   return {
 
@@ -11,16 +11,16 @@ module.exports = function(dbConfig) {
     },
 
     create: function(req, res) {
-      var userData = { 
-        username: req.body.username, 
-        email: req.body.email, 
+      var userData = {
+        username: req.body.username,
+        email: req.body.email,
         password: req.body.hashedPassword
       };
       poetsRepo.create(userData, function(err, user) {
         respond(err, res, function() {
           passport.authenticate('local')(req, res, function() {
             res.redirect('/poem');
-          })
+          });
         });
       });
     }
