+++ server/views/register.jade
